Update MainActivity.java
This is the Latest Android studio code this code is responsible for the following attributes. 
1) This is the Main activity where all of the processing of the project takes place. 
2) Some of the Android studio libraries used for Progress bar, Sending Notification to notification bar and using RetroFit.
3) This code also checks if user's input is a supported bus number else the user is returned an error. 
4) Depending on the Bus's Current Capacity zone, the process bar's color and length changes for visual aid and a notification is also sent to the user.
_____________________________________________________
I tried t include OC Transport's API but got no success. so the destination of the Bus was hard coded in the application.  
__________
Minimum requirement of the application was to display the capacity zone of the bus. 
The Progress Bar, Notification Push,Text/Progress bar's changing colors and Applications Icon were all done to build a good looking and Practical application.